Add simplified conceptual model based on markov chain
The desire based model was not based on actual data and couldn't be validated with actual data either. So it was changed to a Markov-chain based model which can be calibrated from time use data.

Furthermore, the energy impact of activities is difficult to calibrate as well as it would also need a lot of data, so it has been simplified.